If the split prop changes, then discard the size prop and draggedSize state
Scenario: you're horizontally split, you drag a pane to 90% which happens to be 1000px. You change the split to vertical, and now the primary pane is also 1000px, which leaves no room for the secondary pane.
A smarter solution would be to use percentages, but this package doesn't support percentages yet